==> notebooks/01_data_cleaning.ipynb <==
# import
import numpy as np
import pandas as pd
import random
import re
import warnings
warnings.filterwarnings('ignore')

### Create a header and an ID for each file 

admin_header_name = ["Job", "URL", "Company", "Location", "Posted Date", "Classification"]
admin_file_name = 'NZ_Admin_JOBS.xlsx'
admin_id = 0

banking_header_name = ["Job", "URL", "Company", "Location", "Classification_2", "Posted Date", "Classification"]
banking_file_name = 'NZ_Banking_JOBS.xlsx'
banking_id = 1

ceo_header_name = ["Job", "URL", "Company", "Location", "Classification", "Posted Date"]
ceo_file_name = 'NZ_CEO_JOBS.xlsx'
ceo_id = 2

construction_header_name = ["Job", "URL", "Company", "Location", "Classification_2", "Posted Date", "Classification"]
construction_file_name = 'NZ_Construction_JOBS.xlsx'
construction_id = 3

# ----------------------------------Functions----------------------------------

### 1. read_file 
Read the file to dataframe and add a header to it

def read_file(file_name, header_name, file_id):
    # read the file and add a header
    df = pd.read_excel(file_name, header=None, names=header_name, engine='openpyxl')
    if file_id == 1 or file_id == 3:
        df = df.drop('Classification_2',axis=1)
    elif file_id == 2 or file_id == 3:
        df = df.iloc[1:]
    return df

### 2. cut_half 
In the raw data, there exists many dupications. This function is used to cut the duplicated strings half.
E.g., it cuts "TaurangaTauranga" to "Tauranga"

def cut_half(x):
    # a function used in data_cleaning
    # cut the duplicated string half
    # e.g., cut "TaurangaTauranga" to "Tauranga"
    
    if x != None:
        if x[-4:] == "area":
            return (x[0:((len(x)-4)//2)])
        else:
            return (x[0:(len(x)//2)])
    else:
        return None

### 3. find_number 
In the raw data, "salary" and "salary infor" are merged into the same column. This function separates "salary" and "salary info".

def find_number(x):
    # separate "salary" and "salary info"
    if x != None and not pd.isna(x):
        if any(char.isdigit() for char in x):
            return x + ";"
        else:
            return ";" + x 

### 4. low_high_salary
In the raw data, the format of the salary is messy. E.g., it has both per hour and yearly pay in the form of $20, $20k, or $40,000. However, the low and high salary are always separated by "-". This function separates the high and lowest and highest salary, and then we estimate the empty data.

def low_high_salary(x, file_id):
    # refine the salary using regular expression
    # get the lowest and the highest salary
    
    output = []
    if x != None:
        x = x.replace(',','')
        x = ''.join((ch if ch in '0123456789.' else ' ') for ch in x)
        x = x.replace(' .','').replace(' 0','0')
        listOfNumbers = [float(i) for i in x.split()]
        if len(listOfNumbers) == 1:
            output = [listOfNumbers[0], listOfNumbers[0]]
        elif len(listOfNumbers) > 1:
            output = [listOfNumbers[0], listOfNumbers[1]]
        # if "0k " in x or "5k " in x or "0K " in x or "5K " in x:
        if bool(re.match("\dk", x)):
            output = [i * 1000 for i in output]
        if output and output[0] < 1000 and output[1] < 1000:
            output = [i * 1760 for i in output]
        if output and output[0] < 1000 and output[1] > 1000:
            output[0] = output[0] * 1000
    if not output:
        if file_id == 0:
            lo = random.uniform(35000, 55000)
            output = [lo, random.uniform(lo, 55000)]
        elif file_id == 1:
            lo = random.uniform(77000, 141000)
            output = [lo, random.uniform(lo, 141000)]
        elif file_id == 2:
            lo = random.uniform(84000, 255000)
            output = [lo, random.uniform(lo, 255000)]
    listToStr = ' '.join([str(int(i)) for i in output])       
    return listToStr

### 5. fix_posted_data
In the raw data. the "posted data" could take the form of either "featured at" or $4d ago, at$. We refine them all to "days" (integer)

def fix_posted_date(x):
    # refine the posted date
    if "Featured" in x or bool(re.search("\d[h,m]", x)):
        return str(0)
    elif bool(re.search("\dd", x)):
        return ''.join((ch if ch in '0123456789' else '') for ch in x)

### 6. fix_company
In the raw data, sometimes the company has a "at " at the begining. This function fixs this problem. 

def fix_company(x):
    if len(x) > 3 and x[:3] == "at ":
        return x[3:]
    else:
        return x

### 7. The main data preprocessing function

def data_cleaning(df, file_id):
    # data cleaning
    # file_id identifies which file it is
    
    # 1. Refine the "posted date" column
    
    column_date_place = df["Posted Date"].str.split(",", n = -1, expand = True) 
    df["Posted Date"] = column_date_place[0]
    df["Posted Place"] = column_date_place[2]
    
    
    # 2. Refine the "location" column
    
    # 2.1 Some rows contain the salary after "," and we need to remove them
    df["Location"] = df["Location"].str.split(",", n = 1, expand = True)[0]
    
    # 2.2 Need to remove the "location:" at the beginning of the strings, and then separate the region and the city
    column_location = df["Location"].str.split(": ", n = 3, expand = True) 
    df['Region'] = column_location[1]
    df['City'] = column_location[2]
    df = df.drop('Location',axis=1)
    
    # 2.3 Remove duplications
    df["Region"] = df["Region"].apply(cut_half)
    df["City"] = df["City"].apply(cut_half)
    
    
    # 3. Refine "classification"
    # Some rows contain the "classification" and the "subclassification", others contain the salary
    # Analyse "classification" and "salary" separately as df_classification and df_salary
    
    # 3.1 analyse "classification" 
    
    df_classification = df[df.Classification.str.contains(':',case=False)]
    column_classifications = df_classification["Classification"].str.split("subClassification: ", n = -1, expand = True)
    df_classification['Classification'] = column_classifications[0].str.split("classification: ", n = -1, expand = True)[1].apply(cut_half)
    df_classification['Sub-classification'] = column_classifications[1].apply(cut_half)

    
    # 3.2 analyse "salary"
    df_salary = df[~df.Classification.str.contains(':',case=True)]
    df_salary = df_salary.rename(columns = {'Classification':'Salary'})
    
    # 3.3 merge "classification" and "salary"
    df = pd.merge(df_classification, df_salary, how='outer')
    df = df[["Job", "URL", "Company", "Posted Date", "Posted Place", "Classification", "Sub-classification", "Region", "City", "Salary"]]
    
    
    # 4. Refine "salary"
    
    # 4.1 separate into "salary" and "salary info"
    df["Salary"] = df["Salary"].apply(find_number)
    column_salary_info = df["Salary"].str.split(";", n = 2, expand = True) 
    df["Salary"] = column_salary_info[0]
    if column_salary_info.shape[1] != 1:
        df["Salary Info"] = column_salary_info[1]
    else:
        df["Salary Info"] = np.nan
    
    # 4.2 obtain the lowest and the highest salary
    column_low_high = df["Salary"].apply(low_high_salary, args=(file_id,))
    column_low_high = column_low_high.str.split(" ", n = -1, expand = True) 
    df["Lowest Salary"] = column_low_high[0]
    df["Higest Salary"] = column_low_high[1]
    df = df.drop('Salary',axis=1)
    
    
    # 5 Merge "Company" and "Posted Place"
    
    df["Company"] = df["Company"].fillna(df['Posted Place'])
    df = df.drop('Posted Place',axis=1)
    
    
    # 6 Refine the posted date
    df["Posted Date (Days Ago)"] = df["Posted Date"].apply(fix_posted_date)
    df = df.drop('Posted Date',axis=1)
    
    
    # 7 Refine "Company"
    df["Company"] = df["Company"].apply(fix_company)
    
    
    #8 Remove null
    df = df.fillna(value='NO DATA')
    
    return df


def print_info(df):
    print("Total number of jobs: "+str(len(df.index)))
    print("Category of jobs: "+str(len(df.Job.unique())))

def load_col(df, col):
    pd.set_option('display.max_colwidth', None)
    nan_va = df[col].isna().sum()
    print(df[col].head(15))
    print("----------------------------------------------------------")
    print("There are "+ str(nan_va)+" NaN values out of "+str(len(df.index))+" jobs")

# --------------------------------Load data--------------------------------

##### 1. Admin jobs

df_admin = read_file(admin_file_name, admin_header_name, admin_id)

##### 2. Banking jobs

df_banking = read_file(banking_file_name, banking_header_name, banking_id)

##### 3. CEO jobs

df_ceo = read_file(ceo_file_name, ceo_header_name, ceo_id)

#### 4. Construction jobs

df_construction = read_file(construction_file_name, construction_header_name, construction_id)

# --------------------------------The raw data is faulty--------------------------------

### 1. The "Company" column

load_col(df_admin,"Company")

There are 22 NaN values

### 2. The "Location" column

load_col(df_admin,"Location")

1. There is a "Location: " at the beginning of every cell
2. For some rows, the "location" and the "area" are merged together in one cell
3. The location is duplicated
4. In some rows, the payment is also merged into the cell

### 3. The "Posted Date" column

load_col(df_admin,"Posted Date")

1. Some rows have "Featured,at,Private Advertiser"
2. Some rows have "Featured,at,"
3. Some rows have ",Private Advertiser" after the posted date
4. Some are measured by "m" (months) and others are measured by "d" (days)

### 4. The "Classification" column

load_col(df_admin,"Classification")

1. Every cell is merged with "classification" and "subclassification"
2. Every cell starts with "classification: "
3. Every cell has "subClassification: " in the middle
4. Every classification and every subClassification are duplicated

# ----------------------------------Data cleaning----------------------------------

### 1. Admin 

df_admin_final = data_cleaning(df_admin, admin_id)
df_admin_final.to_excel('Cleaned_NZ_Admin_JOBS.xlsx')
df_admin_final.head()

##### Admin job info

print_info(df_admin_final)

### 2. Banking

df_banking_final = data_cleaning(df_banking, banking_id)
df_banking_final.to_excel('Cleaned_NZ_Banking_JOBS.xlsx')
df_banking_final.head()

##### Banking job info

print_info(df_banking_final)

### 3. CEO 

df_ceo_final = data_cleaning(df_ceo, ceo_id)
df_ceo_final.to_excel('Cleaned_NZ_CEO_JOBS.xlsx')
df_ceo_final.head()

##### CEO job info

print_info(df_ceo_final)

### 4. Construction

df_construction_final = data_cleaning(df_construction, construction_id)
df_construction_final.to_excel('Cleaned_NZ_Construction_JOBS.xlsx')
df_construction_final.head()

##### Construction job info

print_info(df_construction_final)
