==> notebooks/02_spark_data_preprocessing.ipynb <==
# Perform data preprocessing using Spark

import findspark
findspark.init()

from pyspark.sql import SparkSession
import pyspark.sql.functions as f
import random
import re

spark = SparkSession \
    .builder \
    .appName("Spark_data_clean") \
    .config("spark.master", "spark://ip-172-31-13-80.ap-southeast-2.compute.internal:7077") \
    .config("spark.jars.packages", "com.crealytics:spark-excel_2.11:0.12.2") \
    .config("spark.executor.cores", "1") \
    .config("spark.executor.num", "1") \
    .getOrCreate()

### 1. Load the raw data and add in the heading

df = spark.read.format("com.crealytics.spark.excel") \
.option("useHeader", "false") \
.option("inferSchema", "true") \
.load("NZ_Admin_JOBS.xlsx")

df = df.withColumnRenamed("_c0", "Job_Title") \
    .withColumnRenamed("_c1", "Web_site") \
    .withColumnRenamed("_c2", "Company_name") \
    .withColumnRenamed("_c3", "Location") \
    .withColumnRenamed("_c4", "Date") \
    .withColumnRenamed("_c5", "Classification")

df.show(5)

### 2. Split the "Location" into "Location" and "Area"

def split_location(location, index):
    if "location: " in location:
        location = location[10:]
    result = location.split("area:")
    if len(result) < 2 and index == 1:
        return None
    else:
        return result[index] 

split_location = f.udf(split_location)
df_location = df.withColumn('loc', split_location(f.col("Location"), f.lit(0))) \
                    .withColumn('area', split_location(f.col("Location"), f.lit(1)))
df_loc = df_location.drop("Web_site").drop("Location").withColumnRenamed("loc","Location")
df_loc.show(5)   

### 3. Remove duplication

def remove_duplicate(x):
    if x is not None:
        x = x.strip()
        x_len = len(x)
        return x[:int(x_len/2)]

remove_duplicate = f.udf(remove_duplicate)
df_location_no_dup = df_loc.withColumn('Location', f.when(f.col('Location').isNotNull(),remove_duplicate(f.col('Location')))) \
                                .withColumn('area', f.when(f.col('area').isNotNull(),remove_duplicate(f.col('area'))))

df_location_no_dup.show(5)

### 4. Fix the “posted date”

from pyspark.sql.functions import regexp_extract
df_date_raw = df_location_no_dup.withColumn('Date', regexp_extract(f.col('Date'), '(\d+[a-z])', 1))
df_date_raw.show(5)

def apply_posted_time(x):
    if 'd' in x:
        return x[:-1]
    if 'm' in x:
        return str(int(x[:-1])*30)
    else:
        return 0
apply_posted_time = f.udf(apply_posted_time)    
df_date = df_date_raw.withColumn("Date", apply_posted_time(f.col("Date")))

df_date.show(5)

### 5. Fix the "Company name"

df_company_name = df_date.withColumn("Company_name", f.when(f.col("Company_name") == "", f.lit("Private Advertiser")).otherwise(f.col("Company_name")))
df_company_name.show(5)

### 6. Fix the "salary"

def get_salary(x):
    if x is not None and any(char.isdigit() for char in x):
        return x
    else:
        return None
get_salary = f.udf(get_salary)

df_salary = df_company_name.withColumn('salary', get_salary(f.col("Classification")))
df_salary.show(15) 

def low_high_salary(x, index):
    # refine the salary using regular expression
    # get the lowest and the highest salary
    
    output = []
    if x != None:
        x = x.replace(',','')
        x = ''.join((ch if ch in '0123456789.' else ' ') for ch in x)
        x = x.replace(' .','').replace(' 0','0')
        listOfNumbers = [float(i) for i in x.split()]
        if len(listOfNumbers) == 1:
            output = [listOfNumbers[0], listOfNumbers[0]]
        elif len(listOfNumbers) > 1:
            output = [listOfNumbers[0], listOfNumbers[1]]
        # if "0k " in x or "5k " in x or "0K " in x or "5K " in x:
        if bool(re.match("\dk", x)):
            output = [i * 1000 for i in output]
        if output and output[0] < 1000 and output[1] < 1000:
            output = [i * 1760 for i in output]
        if output and output[0] < 1000 and output[1] > 1000:
            output[0] = output[0] * 1000
    if not output:
        lo = random.uniform(35000, 55000)
        output = [lo, random.uniform(lo, 55000)]
       
    # listToStr = ' '.join([str(int(i)) for i in output])       
    # return listToStr[index]
    return "%.0f" % output[index]

low_high_salary = f.udf(low_high_salary)
df_low_high_salary = df_salary.withColumn('low_salary', low_high_salary(f.col("salary"), f.lit(0))) \
                    .withColumn('high_salary', low_high_salary(f.col("salary"), f.lit(1)))
df_low_high_salary = df_low_high_salary.drop("salary")
df_low_high_salary.show(15) 

### 7. Fix the classification

def split_classification(x, index):
    if x is not None and "classification: " in x:
        x = x[16:]
        result = x.split("subClassification: ")
        return result[index]
    else:
        return None

split_classification = f.udf(split_classification)
df_classification = df_low_high_salary.withColumn('classification_dup', split_classification(f.col("Classification"), f.lit(0))) \
                    .withColumn('sub-Classification', split_classification(f.col("Classification"), f.lit(1)))
df_classification = df_classification.drop("Classification").withColumnRenamed("classification_dup","Classification")
df_classification.show(5) 

df_classification_no_dup = df_classification.withColumn('Classification', f.when(f.col('Classification').isNotNull(),remove_duplicate(f.col('Classification')))) \
                                .withColumn('sub-Classification', f.when(f.col('sub-Classification').isNotNull(),remove_duplicate(f.col('sub-Classification'))))

df_classification_no_dup.show(5, False)

df_classification_no_dup.repartition(1) \
   .write.format("com.databricks.spark.csv") \
   .option("header", "true") \
   .save("Spark_cleaned_Admin_JOBS.csv")
