==> notebooks/04_regression.ipynb <==
# import
import pandas as pd
import numpy as np

import matplotlib.pyplot as plt
import seaborn as sns
import matplotlib as mpl

import warnings
warnings.filterwarnings("ignore")

from sklearn.preprocessing import LabelEncoder

# **--------------------------PREPARE THE DATA--------------------------**

1. **Load df**

import io
df_admin = pd.read_excel('/content/Cleaned_NZ_Admin_JOBS.xlsx')
df_banking = pd.read_excel('/content/Cleaned_NZ_Banking_JOBS.xlsx')
df_ceo = pd.read_excel('/content/Cleaned_NZ_CEO_JOBS.xlsx')
# Dataset is now stored in a Pandas Dataframe

2. **Add domain for regression**

df_admin['Domain'] = 'Admin'
df_banking['Domain'] = 'Banking'
df_ceo['Domain'] = 'CEO'

# Combine three dataframes
df_all = pd.concat([df_admin, df_banking, df_ceo], ignore_index=True, sort=False)
df_all = df_all[["Job", "Company", "Region", "City", "Lowest Salary", "Higest Salary", "Posted Date (Days Ago)", "Domain"]]
df_all = df_all.dropna()
df_all.head(5)

3. **Set y_df to lower_salary (regression goal)**

y_df = df_all['Lowest Salary']
y_df_encoded = LabelEncoder().fit_transform(y_df.values)

4. **Encode x_label** 

x_df = df_all.apply(LabelEncoder().fit_transform)
x_df.head(5)

5. **Perform train-test split**

from sklearn.model_selection import train_test_split

X_train, X_test, y_train, y_test = train_test_split(x_df, y_df, test_size=0.2, random_state=0)

# **--------------------------EVALUATION FUNCTIONS---------------------**

from sklearn import metrics
from sklearn.model_selection import cross_val_score

def cross_val(model):
    pred = cross_val_score(model, x_df, y_df, cv=5, scoring='r2')
    return pred

def print_evaluate(true, predicted):  
    mae = metrics.mean_absolute_error(true, predicted)
    mse = metrics.mean_squared_error(true, predicted)
    rmse = np.sqrt(metrics.mean_squared_error(true, predicted))
    r2_square = metrics.r2_score(true, predicted)
    print('MAE:', mae)
    print('MSE:', mse)
    print('RMSE:', rmse)
    print('R2 Square', r2_square)
    
def evaluate(true, predicted):
    mae = metrics.mean_absolute_error(true, predicted)
    mse = metrics.mean_squared_error(true, predicted)
    rmse = np.sqrt(metrics.mean_squared_error(true, predicted))
    r2_square = metrics.r2_score(true, predicted)
    return mae, mse, rmse, r2_square

# **-----------------------------REGRESSION WITH XGBOOST-----------------------------**

### **Define a XGBOOST regressor**

import xgboost as xgb
from sklearn.model_selection import RandomizedSearchCV
from scipy.stats import uniform, randint

xgb_model = xgb.XGBRegressor()

params = {
    "learning_rate": uniform(0.01, 0.3), 
    "max_depth": randint(2, 6),
    "n_estimators": randint(100, 1000), 
    "subsample": uniform(0.6, 1)
}

search = RandomizedSearchCV(xgb_model, param_distributions=params, random_state=12, n_iter=200, cv=3, verbose=1, n_jobs=10, return_train_score=False)

search.fit(x_df, y_df_encoded)

search.cv_results_

**Print the top 5 results**

import numpy as np
results = search.cv_results_
for i in range(1, 5):
  candidates = np.flatnonzero(results['rank_test_score'] == i)
  for candidate in candidates:
      print("Model with rank: {0}".format(i))
      print("Mean validation score: {0:.3f} (std: {1:.3f})".format(
            results['mean_test_score'][candidate],
            results['std_test_score'][candidate]))
      print("Parameters: {0}".format(results['params'][candidate]))

### **Print the best params**

params_best = search.best_params_
params_best

### **Load the best params and print the results**

params = {'learning_rate': 0.0989510082720107,
          'max_depth': 4,
          'n_estimators': 333,
          'subsample': 0.6471626631365215}

xgb_model = xgb.XGBRegressor(**params)
xgb_model.fit(X_train, y_train)

text_pred = xgb_model.predict(X_test)
train_pred = xgb_model.predict(X_train)


print('Test set evaluation:\n_____________________________________')
print_evaluate(y_test, text_pred)
print('Train set evaluation:\n_____________________________________')
print_evaluate(y_train, train_pred)
